fix DB int to decimal

+++ db/migrate/20250520092151_create_promotion_codes.rb
@@ -2,7 +2,7 @@ class CreatePromotionCodes < ActiveRecord::Migration[8.0]
   def change
     create_table :promotion_codes do |t|
       t.string :code, null: false
-      t.integer :discount_amount, null: false
+      t.decimal :discount_amount, null: false
       t.boolean :active, null:false, default: true
       t.datetime :expires_at
       t.boolean :used, null: false, default: false
